text-annotation-tool/requirements.txt
dj-database-url
django-heroku
djangorestframework
django-filter
django-widget-tweaks
